feat(validation): add dashboard layout and widget schemas
- Introduced `dashboardLayoutSchema` and `saveDashboardLayoutInput` for validating dashboard layouts and widget instances.
- Added widget size and position schemas to enforce constraints on widget dimensions and placement.
- Enhanced email validation schemas to include sync intervals and additional fields for email accounts and templates.
- Updated tenant validation to allow positive rent values.
- Added a script to reset thread IDs in the email database after removing subject-based thread matching.

+++ packages/shared/src/validation/tenant.ts
@@ -21,8 +21,8 @@ export const createTenantInput = z.object({
   depositPaid: z.boolean().default(false),
   rentStart: z.string(),
   rentEnd: z.string().optional(),
-  coldRent: z.number().int().positive(),
-  warmRent: z.number().int().positive(),
+  coldRent: z.number().positive(),
+  warmRent: z.number().positive(),
   noticePeriodMonths: z.number().int().positive().optional(),
   rentType: z.enum(rentTypeValues).optional(),
 });
